Moved helper functions into non-test code for reuse elsewhere
