+++ .gitignore
@@ -1,2 +1,3 @@
 *.swp
 *.swo
+vendor
